requirements.txt
# ADK agent (keyword_researcher) and Vertex AI Agent Engine deploy
google-adk>=0.1.0
python-dotenv>=1.0.0
pydantic==2.12.2
cloudpickle==3.1.2
google-cloud-aiplatform[adk,agent_engines]>=1.120.0

